- organizamos carpetas y separamos funciones
