+++ client/taekwondo-app-ang/package.json
@@ -28,6 +28,7 @@
     "brackets-model": "^1.5.0",
     "brackets-viewer": "^1.6.2",
     "rxjs": "~7.8.0",
+    "sweetalert2": "^11.23.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.15.0"
   },
