+++ client/taekwondo-app-ang/src/app/core/utils/id.utils.ts
@@ -19,3 +19,6 @@ export function generateId(length: number = 12): string {
 
   return id;
 }
+
+
+
