+++ client/taekwondo-app-ang/src/app/core/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './date.utils';
+export * from './id.utils';
