+++ client/taekwondo-app-ang/src/app/features/combates/combates.component.ts
@@ -2,11 +2,10 @@ import { Component, OnInit, signal } from '@angular/core';
 import { IInfoTorneoCategoria } from '@app/interface/request/info-torneo-categoria';
 import { IGestionTorneoData, ITorneoCreateData } from '@app/interface/torneo-data';
 import { dataTorneo } from '@app/json/torneo';
-import { TournamentService } from '@app/services/tournament-service';
 import { AllSharedImports } from '@app/shared/all-shared-imports';
 
 import { GestionTorneoComponent } from '@app/shared/components/gestion-torneo/gestion-torneo.component';
-import { InputStage } from 'brackets-model';
+
 
 @Component({
   selector: 'app-combates',
