+++ client/taekwondo-app-ang/src/app/services/brackets/viewer-service.ts
