+++ client/taekwondo-app-ang/src/app/services/helper.service.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({ providedIn: 'root' })
+export class HelperService {
+
+
+}
